chore: move tasks to separate file, config cleanup
